feat: add maybeT monadic transformer
Generally, a monadic transformer allows one to merge two monads into
a single tool with the traits of both parent monads. In the case of the
`maybe` monad, the resultant tool will behave with maybe semantics: if
the contained value is null then the `map` function will not be called,
else it will.

When combined with an array (which is monadic in a sense), then the
`maybeT` transform will iterate the `map` function over each element of
the array and call the `map` function only if the element is non-null.

This change adds special handling for `Promises` - or more generically,
thenables - and treats the `then` function as a `map` function. This
isn't technically true, since `then` is actually a `flatMap` function
but these pedantics are relatively minor. When `maybeT` is combined with
`Promise`, the `map` function will be called once the promise resolves
if the resolve value is not nullable.