perf: use rollup.js to bundle lib
By using rollup we can support es module imports for other webpack and
rollup users via the `package.json` `module` field. This allows for
tree-shaking for this lib (especially for something like `MaybeT` which
is not likely used very often).

The other advantage to using rollup is some minor performance boosts and
removal of circular import issues on the consumer. Currently using
maybetyped in a repo with webpack or rollup results in warnings about
circular imports. While this doesn't break anything, it is definitely
sub-optimal to be the source of these warnings.